backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
bcrypt==4.1.3
marshmallow==3.21.3
reportlab==4.2.2
